Separated Main and Registration activities
