+++ app/src/main/AndroidManifest.xml
@@ -20,7 +20,8 @@
         </activity>
         <activity android:name=".LoginActivity" />
         <activity android:name=".HomepageActivity" />
-        <activity android:name=".CreateListingActivity"></activity>
+        <activity android:name=".CreateListingActivity" />
+        <activity android:name=".RegistrationActivity"></activity>
     </application>
 
 </manifest>
