+++ app/src/main/res/layout/activity_main.xml
@@ -5,85 +5,52 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context="com.thewhitewalkers.parkhere.MainActivity">
-<FrameLayout
-    android:layout_width="wrap_content"
-    android:layout_height="495dp"
-    tools:layout_editor_absoluteY="8dp"
-    tools:layout_editor_absoluteX="0dp">
 
-    <LinearLayout
-        android:layout_width="368dp"
-        android:layout_height="wrap_content"
-        android:layout_centerHorizontal="true"
-        android:layout_centerVertical="true"
-        android:orientation="vertical"
+    <FrameLayout
+        android:layout_width="wrap_content"
+        android:layout_height="495dp"
         tools:layout_editor_absoluteX="8dp"
-        tools:layout_editor_absoluteY="106dp">
+        tools:layout_editor_absoluteY="16dp">
 
-        <EditText
-            android:id="@+id/editTextEmail"
-            android:layout_width="337dp"
+        <LinearLayout
+            android:layout_width="368dp"
             android:layout_height="wrap_content"
-            android:layout_margin="15dp"
-            android:hint="Enter your email"
-            android:inputType="textEmailAddress" />
+            android:layout_centerHorizontal="true"
+            android:layout_centerVertical="true"
+            android:orientation="vertical"
+            tools:layout_editor_absoluteX="8dp"
+            tools:layout_editor_absoluteY="106dp"
+            android:weightSum="1">
 
-        <EditText
-            android:id="@+id/editTextPassword"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="15dp"
-            android:hint="Enter your password"
-            android:inputType="textPassword" />
+            <Button
+                android:id="@+id/buttonRegister"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:layout_margin="15dp"
+                android:text="Register" />
 
-        <EditText
-            android:id="@+id/editTextFirstName"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="15dp"
-            android:hint="First Name"
-            android:inputType="text" />
+            <TextView
+                android:id="@+id/textViewParkHere"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:layout_centerHorizontal="true"
+                android:layout_marginTop="30dp"
+                android:text="ParkHere"
 
-        <EditText
-            android:id="@+id/editTextLastName"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="15dp"
-            android:hint="Last Name"
-            android:inputType="text" />
+                android:textAlignment="center"
+                android:textColor="@color/colorBlack"
+                android:textSize="40sp" />
 
-        <EditText
-            android:id="@+id/editTextLocation"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="15dp"
-            android:hint="Location"
-            android:inputType="text" />
+            <Button
+                android:id="@+id/buttonLogin"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:layout_margin="15dp"
+                android:text="Login" />
 
-        <EditText
-            android:id="@+id/editTextPhoneNumber"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_margin="15dp"
-            android:hint="Phone Number"
-            android:inputType="text" />
 
-        <Button
-            android:id="@+id/buttonRegister"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_margin="15dp"
-            android:text="Register User" />
+        </LinearLayout>
+    </FrameLayout>
 
-        <TextView
-            android:id="@+id/textViewSignin"
-            android:textAlignment="center"
-            android:text="Already Registered? Sign in Here"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content" />
-
-    </LinearLayout>
-
-</FrameLayout>
 
 </android.support.constraint.ConstraintLayout>
